1.1.0 release

This releases adds functions to open directories/files and to support VDP
and fixes minor issues

+++ src/default_blockinput.c
@@ -154,7 +154,7 @@ static int _def_read(udfread_block_input *p_gen, uint32_t lba, void *buf, uint32
     pos   = (off_t)lba * UDF_BLOCK_SIZE;
 
     while (got < bytes) {
-        ssize_t ret = pread(p->fd, ((char*)buf) + got, bytes - got, pos + got);
+        ssize_t ret = pread(p->fd, ((char*)buf) + got, bytes - got, pos + (off_t)got);
 
         if (ret <= 0) {
             if (ret < 0 && errno == EINTR) {
@@ -165,7 +165,7 @@ static int _def_read(udfread_block_input *p_gen, uint32_t lba, void *buf, uint32
             }
             break;
         }
-        got += ret;
+        got += (size_t)ret;
     }
 
     return got / UDF_BLOCK_SIZE;
